Fix for inflection when your absolute filepath happens to match `v[0-9]` (#1618)
in the parent structure. This can happen on nix.

For example, with a filepath of:

 /nix/store/xdav9lnh1idxynnqayfr3dclm4i2dmq3-ruby3.2.2-datadog_api_client-2.10.0/lib/ruby/gems/3.2.0/gems/datadog_api_client-2.10.0/lib/datadog_api_client/v1/api/authentication_api.rb

You'll attempt to inflect AuthenticationAPI as AuthenticationApi, as we attempt to
lookup an override of "v9.authentication_api", which does not exist.

(Look closely: `xdav9lnh1idxynnqayfr3dclm4i2dmq3`) has `v9` in it. b13e6dd5419bf1ac1d0d4f74382345842d970af7